2.0.1 (02CD)
Changes:
* Fixed issue where the duration set from an Apple Script would not be
translated into minutes, but seconds.
* Made it possible to set durations with decimals when calling
RocketFuel from an Apple Script.

+++ RocketFuel/AppDelegate.swift
@@ -35,7 +35,7 @@ class AppDelegate: NSObject, NSApplicationDelegate {
         }
     }
 
-    func applicationShouldActivateWithDuration(duration: Int) {
+    func applicationShouldActivateWithDuration(duration: Double) {
         self.loadStatusItemController()
         self.statusItemController.requestActivation(duration)
     }
